+++ cpp/tests/test_utils.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2020, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -45,6 +45,8 @@ extern "C" {
 #include <rmm/rmm.h>
 
 #include "utilities/error_utils.h"
+
+#include "converters/COOtoCSR.cuh"
 
 #ifndef CUDA_RT_CALL
 #define CUDA_RT_CALL(call)                                                               \
@@ -124,7 +126,7 @@ void printv(size_t n, T* vec, int offset)
     dev_ptr + offset,
     dev_ptr + offset + n,
     std::ostream_iterator<T>(
-      std::cout, " "));  // Assume no RMM dependency; TODO: check / test (potential BUG !!!!!)
+      std::cout, " "));  // Assume no RMM dependency; FIXME: check / test (potential BUG !!!!!)
   std::cout << std::endl;
 }
 
@@ -622,8 +624,51 @@ int read_binary_vector(FILE* fpin, int n, std::vector<double>& val)
   return 0;
 }
 
+// FIXME: A similar function could be useful for CSC format
+//        There are functions above that operate coo -> csr and coo->csc
+/**
+ * @tparam
+ */
+template <typename VT, typename ET, typename WT>
+std::unique_ptr<cugraph::experimental::GraphCSR<VT, ET, WT>> generate_graph_csr_from_mm(bool& directed, std::string mm_file)
+{
+  VT number_of_vertices;
+  ET number_of_edges;
+  
+  FILE* fpin = fopen(mm_file.c_str(), "r");
+  EXPECT_NE(fpin, nullptr);
+
+  VT number_of_columns = 0;
+  MM_typecode mm_typecode{0};
+  EXPECT_EQ(mm_properties<VT>(
+                              fpin, 1, &mm_typecode, &number_of_vertices, &number_of_columns, &number_of_edges),
+            0);
+  EXPECT_TRUE(mm_is_matrix(mm_typecode));
+  EXPECT_TRUE(mm_is_coordinate(mm_typecode));
+  EXPECT_FALSE(mm_is_complex(mm_typecode));
+  EXPECT_FALSE(mm_is_skew(mm_typecode));
+
+  directed = !mm_is_symmetric(mm_typecode);
+
+  // Allocate memory on host
+  std::vector<VT> coo_row_ind(number_of_edges);
+  std::vector<VT> coo_col_ind(number_of_edges);
+  std::vector<WT> coo_val(number_of_edges);
+
+  // Read
+  EXPECT_EQ((mm_to_coo<VT, WT>(
+                               fpin, 1, number_of_edges, &coo_row_ind[0], &coo_col_ind[0], &coo_val[0], NULL)),
+            0);
+  EXPECT_EQ(fclose(fpin), 0);
+
+  cugraph::experimental::GraphCOOView<VT, ET, WT> cooview(
+    &coo_col_ind[0], &coo_row_ind[0], &coo_val[0], number_of_vertices, number_of_edges);
+
+  return cugraph::coo_to_csr(cooview);
+}
+
 ////////////////////////////////////////////////////////////////////////////////
-// TODO: move this code to rapids-core
+// FIXME: move this code to rapids-core
 ////////////////////////////////////////////////////////////////////////////////
 
 // Define RAPIDS_DATASET_ROOT_DIR using a preprocessor variable to