+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 - PR #756 Add Force Atlas 2 layout
 - PR #822 Added new functions in python graph class, similar to networkx
 - PR #840 OPG degree
+- PR #881 Raft integration infrastructure
 - PR #875 UVM notebook 
 
+
 ## Improvements
+- PR #882 Add Force Atlas 2 to benchmarks
+- PR #876 Add BFS C++ tests
 - PR #817 Add native Betweenness Centrality with sources subset
 - PR #764 Updated sssp and bfs with GraphCSR, removed gdf_column, added nullptr weights test for sssp
 - PR #765 Remove gdf_column from connected components